refactor: project publish
- project publish 로직 컨트롤러에서 서비스로 이동
- publishProject()에서 update 대신 save를 사용(update는 캐시 로직 때문에 사용이 어려움)
- game repository에서 game을 찾을 때 user id도 조회하도록 쿼리 수정(gameService.updateGame()에서 사용)
