+++ src/game/game.repository.ts
@@ -59,6 +59,7 @@ export class GameRepository extends Repository<Game> {
       .leftJoin('game.likes', 'likes')
       .leftJoin('game.user', 'user')
       .addSelect('user.nickname')
+      .addSelect('user.id')
       .addSelect('likes.email')
       .loadRelationCountAndMap('game.likeCount', 'game.likes')
       .where('game.id = :id', { id: id })
@@ -71,6 +72,7 @@ export class GameRepository extends Repository<Game> {
       .leftJoin('game.likes', 'likes')
       .leftJoin('game.user', 'user')
       .addSelect('user.nickname')
+      .addSelect('user.id')
       .addSelect('likes.email')
       .limit(limit)
       .offset(offset)
